refactor: move hover persiting code into "stack" from "trigger"
