+++ packages/overlay/src/OverlayTrigger.ts
@@ -309,8 +309,6 @@ export class OverlayTrigger extends SpectrumElement {
             case 'click':
                 if (this.clickContent) {
                     this.open = event.type;
-                } else if (this.closeHoverOverlay) {
-                    event.preventDefault();
                 }
                 return;
             case 'longpress':
