+++ packages/overlay/test/overlay-trigger-hover-click.test.ts
@@ -9,7 +9,13 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import { expect, fixture, html, oneEvent } from '@open-wc/testing';
+import {
+    elementUpdated,
+    expect,
+    fixture,
+    html,
+    oneEvent,
+} from '@open-wc/testing';
 import '@spectrum-web-components/popover/sp-popover.js';
 import '@spectrum-web-components/action-button/sp-action-button.js';
 import '@spectrum-web-components/button/sp-button.js';
@@ -21,6 +27,7 @@ import { TriggerInteractions } from '@spectrum-web-components/overlay/src/overla
 import '@spectrum-web-components/overlay/overlay-trigger.js';
 import { ActionButton } from '@spectrum-web-components/action-button';
 import { sendMouse } from '../../../test/plugins/browser.js';
+import { clickAndHoverTargets } from '../stories/overlay.stories.js';
 
 describe('Overlay Trigger - Hover and Click', () => {
     it('toggles open and closed on click', async () => {
@@ -109,5 +116,83 @@ describe('Overlay Trigger - Hover and Click', () => {
 
             expect(interaction).equals('click');
         }
+    });
+    it('persists a hover overlay when clicking its trigger and closes the next highest overlay on the stack', async () => {
+        const root = document.createElement('div');
+        root.style.width = '100vw';
+        root.style.height = '100vh';
+        root.style.display = 'grid';
+        root.style.placeContent = 'center';
+        const test = await fixture(clickAndHoverTargets(), {
+            parentNode: root,
+        });
+
+        const overlayTrigger1 = test.querySelector(
+            'overlay-trigger[placement="right"]'
+        ) as OverlayTrigger;
+        const overlayTrigger2 = test.querySelector(
+            'overlay-trigger[placement="left"]'
+        ) as OverlayTrigger;
+
+        await elementUpdated(overlayTrigger1);
+        await elementUpdated(overlayTrigger2);
+
+        const trigger1 = overlayTrigger1.querySelector(
+            '.friendly-target'
+        ) as HTMLButtonElement;
+        const trigger2 = overlayTrigger2.querySelector(
+            '.friendly-target'
+        ) as HTMLButtonElement;
+        const rect1 = trigger1.getBoundingClientRect();
+        const rect2 = trigger2.getBoundingClientRect();
+        let opened = oneEvent(trigger1, 'sp-opened');
+        sendMouse({
+            steps: [
+                {
+                    type: 'move',
+                    position: [
+                        rect1.left + rect1.width / 2,
+                        rect1.top + rect1.height / 2,
+                    ],
+                },
+                {
+                    type: 'click',
+                    position: [
+                        rect1.left + rect1.width / 2,
+                        rect1.top + rect1.height / 2,
+                    ],
+                },
+            ],
+        });
+        await opened;
+        await elementUpdated(overlayTrigger1);
+
+        expect(overlayTrigger1.open).to.equal('click');
+        expect(overlayTrigger2.open).to.undefined;
+
+        opened = oneEvent(trigger2, 'sp-opened');
+        trigger2.focus();
+        await opened;
+        await elementUpdated(overlayTrigger2);
+
+        expect(overlayTrigger1.open).to.equal('click');
+        expect(overlayTrigger2.open).to.equal('hover');
+
+        const closed = oneEvent(trigger1, 'sp-closed');
+        sendMouse({
+            steps: [
+                {
+                    type: 'click',
+                    position: [
+                        rect2.left + rect2.width / 2,
+                        rect2.top + rect2.height / 2,
+                    ],
+                },
+            ],
+        });
+        await closed;
+
+        expect(overlayTrigger1.open).to.be.null;
+        expect(overlayTrigger2.open).to.equal('hover');
     });
 });
